Super minor (but not worthless) code tweaks.

+++ ros/src/inertial_sense_ros.cpp
@@ -2201,7 +2201,7 @@ ros::Time InertialSenseROS::ros_time_from_tow(const double tow)
 
 double InertialSenseROS::tow_from_ros_time(const ros::Time &rt)
 {
-    return (rt.sec - UNIX_TO_GPS_OFFSET - GPS_week_ * 604800) + rt.nsec * 1.0e-9;
+    return ((uint64_t)rt.sec - UNIX_TO_GPS_OFFSET - GPS_week_ * 604800) + rt.nsec * 1.0e-9;
 }
 
 ros::Time InertialSenseROS::ros_time_from_gtime(const uint64_t sec, double subsec)
